fix(cgen/rtl): 修复 x64 下 Variant→LongPtr 指针截断与 StdPicture 赋值崩溃
- Fix 133x: Declare ByRef LongPtr 形参接收 Variant 实参时临时槽用 intptr_t
  (原 int32_t 截断 64 位 GDI+ Font 指针) → 新增 vb6_VariantSetI8/VT_I8 回写,
  vb6_VariantToLongPtr 补 VT_I8/VT_UI8 分支
- Fix 092r: ByVal LongPtr/ULong 实参用 vb6_VariantToLongPtr (原 VariantToLong 截断)
- Fix 110z8/110z9: Set ctrl.Picture = <StdPicture 变量> 走 vb6_SetControlPictureFromCom,
  RHS 直接用变量名避免 emitExpr 误生成 .Width 默认属性
- 测试: 新增 tests/Charts 2020 (LaVolpe 图表组件集), res/redim 等回归 PASS=85

+++ src/backend/cgen_base_type.cpp
@@ -32,7 +32,12 @@ std::string CCodeGen::mapType(Vb6Type type) const {
         case Vb6Type::LongPtr: cType = "intptr_t"; break;   // Fix 081e: architecture-width integer
         case Vb6Type::Single:   cType = "float"; break;
         case Vb6Type::Double:   cType = "double"; break;
-        case Vb6Type::Currency: cType = "int64_t"; break;   // scaled integer
+        // Fix 126: Currency = 64bit/10000 (4 位小数) 的**值** —— 与 Date 一样按值语义
+        // 映射为 double。此前映射 int64_t(存放大整数), 而表达式与 VARIANT 打包都按
+        // 值使用, 于是 `yRange = yRange + CCur(step)` 每步被放大 10000 倍
+        // (Charts 2020 坐标轴数字与柱高严重不符)。放大整数只存在于 VT_CY 变体里,
+        // 由读取侧除回来 (rtl: vb6_VariantToDouble / vb6_Format 的 vtCurrency 分支)。
+        case Vb6Type::Currency: cType = "double"; break;
         case Vb6Type::Date:     cType = "double"; break;    // OLE date
         case Vb6Type::String:   cType = "BSTR"; break;      // wchar_t* wrapper
         case Vb6Type::Object:   cType = "void*"; break;     // IDispatch* → void* for now
@@ -71,7 +76,7 @@ std::string CCodeGen::mapComType(Vb6Type type) const {
         case Vb6Type::Long:     return "int32_t";
         case Vb6Type::LongPtr: return "intptr_t";  // Fix 081e        case Vb6Type::Single:   return "float";
         case Vb6Type::Double:   return "double";
-        case Vb6Type::Currency: return "int64_t";
+        case Vb6Type::Currency: return "double";   // Fix 126: 值语义
         case Vb6Type::Date:     return "double";
         case Vb6Type::String:   return "BSTR";
         case Vb6Type::Object:   return "IUnknown*";
